my-app/src/backend/requirements.txt
langchain
transformers
torch
matplotlib
plotly
fastapi
uvicorn
langchain_community
torch
python-dotenv